Revert "Add refreshToken"
