+++ CryptoBankVerticalWebApi/Database/ApplicationDbContext.cs
@@ -1,5 +1,4 @@
 ﻿using CryptoBankVerticalWebApi.Features.Accounts.Domain;
-using CryptoBankVerticalWebApi.Features.Auth.Model;
 using CryptoBankVerticalWebApi.Features.Users.Domain;
 using Microsoft.EntityFrameworkCore;
 
@@ -10,7 +9,6 @@ namespace CryptoBankVerticalWebApi.Database
         public DbSet<User> Users { get; set; }
         public DbSet<Account> Accounts { get; set; }
         public DbSet<Role> Roles { get; set; }
-        public DbSet<RefreshToken> RefreshTokens { get; set; }
         public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options){ }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
@@ -19,7 +17,6 @@ namespace CryptoBankVerticalWebApi.Database
             MapUsers(modelBuilder);
             MapAccounts(modelBuilder);
             MapRoles(modelBuilder);
-            MapRefreshTokens(modelBuilder);
         }
 
         private void MapUsers(ModelBuilder modelBuilder)
@@ -105,26 +102,6 @@ namespace CryptoBankVerticalWebApi.Database
                     .WithMany(a => a.Accounts)
                     .HasForeignKey(a => a.UserId)
                     .OnDelete(DeleteBehavior.Cascade);
-            });
-        }
-
-        private void MapRefreshTokens(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<RefreshToken>(refreshToken =>
-            {
-                refreshToken.HasKey(r => r.Id);
-
-                refreshToken.Property(e => e.Token)
-              .IsRequired()
-              .HasMaxLength(1000);
-
-                refreshToken.Property(r => r.userId)
-                .IsRequired();
-
-                refreshToken.HasOne(d => d.User)
-                .WithMany(p => p.RefreshTokens)
-                .HasForeignKey(d => d.userId)
-                .OnDelete(DeleteBehavior.ClientSetNull);
             });
         }
     }
