--- CryptoBankVerticalWebApi/Features/Auth/Options/RefreshTokenOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace CryptoBankVerticalWebApi.Features.Auth.Options;
-
-public class RefreshTokenOptions
-{
-    public TimeSpan RefreshTokenExpiration { get; set; }
-}
