+++ CryptoBankVerticalWebApi/Features/Auth/Requests/LoginUser.cs
@@ -14,7 +14,7 @@ namespace CryptoBankVerticalWebApi.Features.Auth.Requests
     public static class LoginUser
     {
         public record Request(LoginModel loginModel) : IRequest<Response>;
-        public record Response(string jwt,string refreshToken);
+        public record Response(string jwt);
 
         public class RequestValidator : AbstractValidator<Request>
         {
@@ -51,10 +51,7 @@ namespace CryptoBankVerticalWebApi.Features.Auth.Requests
 
             public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
             {
-                var user = await _applicationDbContext.Users
-                    .Include(u => u.Roles)
-                    .Include(u=>u.RefreshTokens)
-                    .SingleOrDefaultAsync(u => u.Email==request.loginModel.Email, cancellationToken);
+                var user = await _applicationDbContext.Users.Include(u => u.Roles).SingleOrDefaultAsync(u => u.Email==request.loginModel.Email, cancellationToken);
                 if (user==null)
                 {
                     throw new Exception("Invalid credentials");
@@ -83,10 +80,8 @@ namespace CryptoBankVerticalWebApi.Features.Auth.Requests
                     throw new Exception("Invalid credentials");
                 }
 
-                var (jwt, refreshToken) = await _tokenGenerateService.GenerateTokensAsync(user,cancellationToken);
-           
-
-                return new Response(jwt,refreshToken);
+                var jwt = await _tokenGenerateService.GenerateToken(user);
+                return new Response(jwt);
             }
         }
     }
