+++ CryptoBankVerticalWebApi/Features/Auth/Services/TokenGenerateService.cs
@@ -1,79 +1,48 @@
-﻿using CryptoBankVerticalWebApi.Database;
-using CryptoBankVerticalWebApi.Features.Auth;
-using CryptoBankVerticalWebApi.Features.Auth.Model;
-using CryptoBankVerticalWebApi.Features.Auth.Options;
+﻿using CryptoBankVerticalWebApi.Features.Auth;
 using CryptoBankVerticalWebApi.Features.Users.Domain;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
-using System.Security.Cryptography;
 
 namespace CryptoBankVerticalWebApi.Features.Auth.Services
 {
     public class TokenGenerateService
     {
         private readonly AuthOptions _authOptions;
-        private readonly RefreshTokenOptions _refreshTokenOptions;
-        private readonly ApplicationDbContext _applicationDbContext;
-        private readonly TokenHelper _tokenHelper;
 
-        public TokenGenerateService(IOptions<AuthOptions> authOptions
-            ,IOptions<RefreshTokenOptions> refreshTokenOptions
-            ,ApplicationDbContext applicationDbContext
-            ,TokenHelper tokenHelper) 
+        public TokenGenerateService(IOptions<AuthOptions> authOptions) 
         {
             _authOptions=authOptions.Value;
-            _refreshTokenOptions=refreshTokenOptions.Value;
-            _applicationDbContext=applicationDbContext;
-            _tokenHelper=tokenHelper;
         }
-        public async Task<(string accessToken,string refreshToken)> GenerateTokensAsync(User user,CancellationToken cancellationToken)
+        public async Task<string> GenerateToken(User user)
         {
-            var refreshTokens = user.RefreshTokens
-                .OrderByDescending(x => x.CreatedAt)
-                .ToArray();
-
-            var accessToken = _tokenHelper.GenerateAccesToken(user);
-            var refreshToken = await _tokenHelper.GenerateRefreshToken();
-
-            await using var transaction = await _applicationDbContext.Database.BeginTransactionAsync(cancellationToken);
+            //var role = (UserRole)user.Role;
+            //var roleString = role.ToString();
+            var claims = new List<Claim>
             {
-                try
-                {
-                    var newRefreshToken = new RefreshToken
-                    {
-                        userId = user.Id,
-                        Token = refreshToken,
-                        ExpiryDate = DateTime.Now.Add(_refreshTokenOptions.RefreshTokenExpiration).ToUniversalTime(),
-                        CreatedAt = DateTime.Now.ToUniversalTime(),
-                    };
+                new Claim(ClaimTypes.NameIdentifier,user.Id.ToString()),
+                new Claim(ClaimTypes.Email,user.Email),
+            };
 
-                    user.RefreshTokens.Add(newRefreshToken);
-                    _applicationDbContext.Add(newRefreshToken);
-                    await _applicationDbContext.SaveChangesAsync(cancellationToken);
-
-                    var NotRevokeRefreshToken = refreshTokens.FirstOrDefault(t => !t.Revoke);
-                    if (NotRevokeRefreshToken!=null)
-                    {
-                        NotRevokeRefreshToken.Revoke = true;
-                        NotRevokeRefreshToken.ReplacedByNextToken=newRefreshToken.Id;
-                    }
-
-                    var overdueTokens = refreshTokens.Where(x => x.ExpiryDate<=DateTime.Now.ToUniversalTime()).ToArray();
-
-                    _applicationDbContext.RefreshTokens.RemoveRange(overdueTokens);
-
-                    await _applicationDbContext.SaveChangesAsync(cancellationToken);
-
-                    await transaction.CommitAsync(cancellationToken);
-                }
-                catch (Exception)
-                {
-                    await transaction.RollbackAsync(cancellationToken);
-                }
-                return (accessToken, refreshToken);
+            foreach (var role in user.Roles)
+            {
+                claims.Add(new Claim(ClaimTypes.Role, role.Name.ToString()));
             }
+
+            var keyBytes = Convert.FromBase64String(_authOptions.Jwt.SigningKey);
+            var key = new SymmetricSecurityKey(keyBytes);
+            var credentials = new SigningCredentials(key, SecurityAlgorithms.HmacSha256Signature);
+            var expires = DateTime.Now + _authOptions.Jwt.Expiration;
+            var token = new JwtSecurityToken(
+                _authOptions.Jwt.Issuer,
+                _authOptions.Jwt.Audience,
+                claims: claims,
+                expires: expires,
+                signingCredentials: credentials
+            );
+            var tokenHandler = new JwtSecurityTokenHandler();
+            return new JwtSecurityTokenHandler().WriteToken(token);
         }
     }
 }
