+++ CryptoBankVerticalWebApi/Features/Users/Domain/User.cs
@@ -1,5 +1,4 @@
 ﻿using CryptoBankVerticalWebApi.Features.Accounts.Domain;
-using CryptoBankVerticalWebApi.Features.Auth.Model;
 
 namespace CryptoBankVerticalWebApi.Features.Users.Domain
 {
@@ -9,7 +8,6 @@ namespace CryptoBankVerticalWebApi.Features.Users.Domain
         {
             Accounts = new HashSet<Account>();
             Roles = new HashSet<Role>();
-            RefreshTokens = new HashSet<RefreshToken>();
         }
         public long Id { get; set; }
         public string Email { get; set; } = string.Empty;
@@ -21,7 +19,6 @@ namespace CryptoBankVerticalWebApi.Features.Users.Domain
         public DateTime CreatedAt { get; set; } = DateTime.Now.ToUniversalTime();
         public DateTime? UpdatedAt { get; set; }
 
-        public virtual ICollection<RefreshToken> RefreshTokens { get; set; }
         public virtual ICollection<Account> Accounts { get; set; }
         public virtual ICollection<Role> Roles { get; set; }
     }
