+++ CryptoBankVerticalWebApi/appsettings.json
@@ -13,8 +13,7 @@
         "Issuer": "CryptoBankWebApi",
         "Audience": "CryptoBankWebApi",
         "Expiration": "00:05:00"
-      },
-      "RefreshTokenExpiration": "24:00:00"
+      }
     },
     "Accounts": {
       "AccountOptions": {
